requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
langchain==0.1.4
langchain-community==0.0.20
langchain-openai==0.0.5
tiktoken==0.5.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2
pgvector==0.2.3
supabase==2.3.4
